Safari-611.1.10/LayoutTests/fast/line-grid/line-snap-parsing-expected.txt

Test parsing of the CSS line-snap property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("-webkit-line-snap: none") is "none"
PASS test("-webkit-line-snap: baseline") is "baseline"
PASS test("-webkit-line-snap: contain") is "contain"
PASS test("-webkit-line-snap: ;") is ""
PASS test("-webkit-line-snap: 1") is ""
PASS test("-webkit-line-snap: 1.2") is ""
PASS test("-webkit-line-snap: -1") is ""
PASS test("-webkit-line-snap: 12px") is ""
PASS testComputedStyle("none") is "none"
PASS testComputedStyle("") is "none"
PASS testComputedStyle("12px") is "none"
PASS testInherited("none") is "none"
PASS testInherited("baseline") is "baseline"
PASS testInherited("contain") is "contain"
PASS successfullyParsed is true

TEST COMPLETE
